requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
pydantic
passlib
PyJWT>=2.8